adapt to v2 and add example
